Added account image update logic

+++ app/Actions/Course/UpdateCourseImageAction.php
@@ -5,7 +5,6 @@ namespace App\Actions\Course;
 use App\Data\Course\Requests\UpdateCourseImageData;
 use App\Models\Course;
 use Exception;
-use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use Throwable;
 
@@ -17,25 +16,24 @@ class UpdateCourseImageAction
      * @param UpdateCourseImageData $courseImageData
      * @param Course $course
      * @return Course
-     * @throws Throwable
+     * @throws Exception
      */
     public function handle(UpdateCourseImageData $courseImageData, Course $course): Course
     {
-        return DB::transaction(function () use ($courseImageData, $course) {
-            try {
-                $oldPath = $course->image_path;
-                $newPath = $courseImageData->image->store('/', 'courses');
-                $course->setImage($newPath)->save();
-                if ($oldPath) {
-                    Storage::disk('courses')->delete($oldPath);
-                }
-                return $course;
-            } catch (Exception $e) {
-                if (isset($newPath)) {
-                    Storage::disk('courses')->delete($newPath);
-                }
-                throw $e;
-            }
-        });
+        $oldPath = $course->image_path;
+        $newPath = $courseImageData->image->store('/', 'courses');
+
+        try {
+            $course->setImage($newPath)->save();
+        } catch (Exception $e) {
+            Storage::disk('courses')->delete($newPath);
+            throw $e;
+        }
+
+        if ($oldPath && Storage::disk('courses')->exists($oldPath)) {
+            Storage::disk('courses')->delete($oldPath);
+        }
+
+        return $course;
     }
 }
